+++ CHANGELOG.md
@@ -26,7 +26,16 @@ Additional labels for pre-release metadata:
 * beta.x: shipped version under testing.
 * rc.x: stable release candidate.
 
-0.2.4 - unreleased
+0.2.5-rc1 - unreleased
+------------------
+
+Added
+* get_account_contract host function
+* is_callable host function
+
+
+
+0.2.4-rc1 - unreleased
 ------------------
 
 Changed
@@ -34,7 +43,7 @@ Changed
 * Signed blocks
 
 Added
-* Bulk transaction
+* Bulk transactions
 * bootstrap from wasm or wasm+txs
 * config from bootstrap
 * network name from bootstrap hash